Break out reusable part of EadSync class to its own thing
This will encapsulate migrating a doc unit from one location to
another, including its PID, access controls, user-generated content
etc
